etl_agent/requirements.txt
# Dependencies for the ETL Agent
python-dotenv
pyodbc
psycopg2-binary
